Рефакторинг
- Пофисксил ошибки с IFormFile
- Добавил summary
- Вынес загрузку FilesStorageConfig в отдельный definition
