+++ backend/Money.Api/Controllers/FileController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Money.Api.Dto.Files;
-using Money.Business;
 using Money.Business.Services;
 using OpenIddict.Validation.AspNetCore;
 
@@ -12,10 +11,16 @@ namespace Money.Api.Controllers;
 [Route("[controller]")]
 public class FileController(FileService fileService) : ControllerBase
 {
+    /// <summary>
+    ///     Загрузить файл.
+    /// </summary>
+    /// <param name="file">Файл для загрузки</param>
+    /// <param name="cancellationToken">Токен отмены запроса.</param>
+    /// <returns></returns>
     [HttpPost]
-    public async Task<FileDto> UploadFile(IFormFile file)
+    public async Task<FileDto> UploadFile(IFormFile file, CancellationToken cancellationToken)
     {
-        var uploadedFile = await fileService.Upload(file);
+        var uploadedFile = await fileService.Upload(file, cancellationToken);
         return FileDto.FromBusinessModel(uploadedFile);
     }
 }
