+++ backend/Money.Api/Definitions/CommonDefinition.cs
@@ -1,5 +1,4 @@
 ﻿using Money.Api.Definitions.Base;
-using Money.Business.Configs;
 
 namespace Money.Api.Definitions;
 
@@ -12,22 +11,6 @@ public class CommonDefinition : AppDefinition
         builder.Services.AddHttpContextAccessor();
         builder.Services.AddResponseCaching();
         builder.Services.AddMemoryCache();
-
-        var filesStorage = builder.Configuration.GetSection("FilesStorage");
-        
-        var filesStorageConfig = filesStorage.Get<FilesStorageConfig>();
-        
-        if (filesStorageConfig == null || string.IsNullOrEmpty(filesStorageConfig.Path))
-        {
-            throw new ApplicationException("FilesStoragePath is missing");
-        }
-
-        if (!Directory.Exists(filesStorageConfig.Path))
-        {
-            Directory.CreateDirectory(filesStorageConfig.Path);
-        }
-        
-        builder.Services.Configure<FilesStorageConfig>(filesStorage);
     }
 
     public override void ConfigureApplication(WebApplication app)
