+++ backend/Money.Business/Money.Business.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="6.0.0-preview.4.21253.5" />
         <PackageReference Include="Microsoft.Extensions.Features" Version="8.0.8" />
         <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.1.0" />
         <PackageReference Include="OpenIddict.Abstractions" Version="5.8.0" />
